sdmmc: upload the files I forgot to upload

+++ drivers/blk/sdmmc/src/sddf_blk/mod.rs
@@ -8,7 +8,7 @@ extern "C" {
         code: *mut BlkOp,
         io_or_offset: *mut u64,
         block_number: *mut u32,
-        count: *mut u16,
+        count: *mut u32,
         id: *mut u32,
     );
 }
@@ -31,7 +31,7 @@ pub struct BlkRequest {
     pub request_code: BlkOp,
     pub io_or_offset: u64,
     pub block_number: u32,
-    pub count: u16,
+    pub count: u32,
     // I suggest use u32 here and change the count to use u32 in sddf_blk
     pub success_count: u32,
     pub count_to_do: u32,
